refactor: split Operator design system into discrete modules + CSS modules
Decompose components/Operator/index.tsx and the 1918-line pages/index.tsx
into a proper module surface:

components/Operator/
  tokens.ts         OP palette + font constants
  hooks.ts          useTicker, useUtcClock, useUptime, useTypewriter,
                    useMouseSpotlight, useReveal, useGsapReady
  data.ts           PLATFORM_TOOLS, HERO_ICONS, TOPICS
  primitives.tsx    Cursor, Prompt, Sec, Role
  Marquee.tsx       continuous ticker
  PlatformBg.tsx    CNCF tools drifting wallpaper
  HeroIconRain.tsx  simpleicons CDN rain
  Topology.tsx      kubectl-get-topology SVG with packet animation
  ClusterGrid.tsx   pod lifecycle simulator
  TalkPreview.tsx   CRT-frame preview card + Talk type
  Header.tsx        sticky operator header with nav
  Footer.tsx        contact footer
  Layout.tsx        OperatorPage wrapper (Head + Scripts + chrome)
  index.tsx         pure re-export surface (24 lines)
  styles.module.css keyframes + utility classes

pages/index.tsx 1918 → 941 lines (-51%): imports components from
~/components/Operator, keeps only home-specific data (TALKS, STACK,
CTAS, TOPBAR_LINKS, CONTACTS) and composition JSX. Home-specific
keyframes + responsive media queries moved to pages/home.module.css.

Inner pages (about/doctrine/writing/talks/presentations/projects/
timeline) switched from namespace `Operator.*` import to direct named
imports for tree-shakeability and clearer call sites.

New useGsapReady hook centralizes the GSAP retry pattern (defensive
boot loop until window.gsap is available), used by HeroIconRain,
Topology, ClusterGrid, and HomePage's scroll-triggered effects.

Lighthouse spot-check after refactor: HOME perf 75 a11y 96 bp 92 seo
100; /about perf 60 a11y 100 bp 92 seo 100 — no regression.
